requirements.txt
streamlit
feedparser
beautifulsoup4
pillow
streamlit-autorefresh

